(Partially) Forward port revision 13803
Apply patch for #2674 by mba

If the gid is the same as a gid of a group with the same name,
set primaryGroup to 0, i.e. "automatic".
If "automatic" was chosen as a group. See if there is group with the
same name as the uid, and if so, set the person's gid to that gid.

git-svn-id: https://oss.gonicus.de/repositories/gosa/branches/2.6-lhm/trunk@13803 594d385d-05f5-0310-b6e9-bd551577e9d8

Conflicts:

    gosa-core/plugins/personal/posix/class_posixAccount.inc