Forward port revision 16310, 17458 and 17925
Apply fix for #2994

Do not check if gosaApplicationFlags changed as it appears
to be not neccessary and causes bugs

Followup commit for #2994

When extracting (key, value) tuples for application parameters
from either an application or a meno groups use a limit in the
split/explode command, so that only the first
double-colon is actually used to split the string. Fixes a bug
where strings containing double-colons get truncated.

Apply fix for #2994

Allow groups to store application flags for associated applications.
They are stored as a special gosaApplicationParameter and only if
they changed compared to the original application (otherwise the gosaApplicationFlag
attribute of the group is removed from LDAP)