Forward port revision 19552 and 16619
Implement a debconf configuration overview (Trac #2025)

- When editting a package list enable the user to
click 'Edit configurations' and get an overview of
and edit all debconf configurations in the given package list.
- Add the possibility to add debconf settings for settings
which are not known to gosa-si.
- Add some safety measures to the debconfTemplate class
to avoid breaking values because of quotes etc.

Apply patch for #2025

Do not check ACLs in the faiDebconfConfigurations class. Instead, do
not display the 'Edit Debconf Configurations' button if the
FAIdebconfInfo ACL is not present for writing, in line with the
faiPackageConfiguration class.

Update button style to match new gosa button style

Use sortableListing in faiDebconfConfigurations

- Update fai debconf configurations overview to use the new class
  sortableListing instead of the old divlist class
- Small bugfix: get_ou now requires a class parameter