Forward port various revisions for Trac #1960
Allow the specification of hidden shares (indicated
by a $ _at the end_ of the share name) by relaxing the
check weither a share name is valid or not

Fix regular expression for the check, weither $
is at the end of the share name or not, so that
foo$bar leads to an error.

Really fix the regular expression for testing of share-
names and change the word of the error message

Check volume for possibly hidden share names, instead
of the name.

Disallow $ in share _names_ again

Followup Commit for r14465:
Change regular expression in error message, too.

Make error message more specific on where the error is.

Check of sharename should still check if only
the allowed chars are used.