Forward port revision 15307
Fix for #4146

Fix wrong condition in check weither a certain ogroup has
a valid driver set or should show "unknown". It tested on
"if attribute gotoXDriver is not empty" while it should
have tested for "if attribute is empty" which caused
unknown to be shown errornous in some cases.

git-svn-id: https://oss.gonicus.de/repositories/gosa/branches/2.6-lhm/trunk@15307 594d385d-05f5-0310-b6e9-bd551577e9d8

Conflicts:

    gosa-plugins/goto/admin/systems/goto/class_workstationService.inc