Forward port revision 19720
When moving or renaming devices, check whether they are in use by a
user, group or object group. If so, display an error message stating
the user. As the reference is not being updated in the user object,
the device would get deleted from the user next time when they are
saved.

git-svn-id: https://oss.gonicus.de/repositories/gosa/branches/2.6-lhm/trunk@19720 594d385d-05f5-0310-b6e9-bd551577e9d8

Conflicts:

    gosa-plugins/goto/admin/devices/class_deviceGeneric.inc