Forward port revision 19650
Added support for optionally taking along members when copying object
groups. This is disabled for workstations, terminals and servers,
which must not belong to more than one object group. A note has been
added to warn users that adding members of a different type will
delete the copied settings.

git-svn-id: https://oss.gonicus.de/repositories/gosa/branches/2.6-lhm/trunk@19650 594d385d-05f5-0310-b6e9-bd551577e9d8

Conflicts:

    gosa-core/locale/core/de/LC_MESSAGES/messages.po
    gosa-core/locale/core/messages.po
    gosa-core/plugins/admin/ogroups/class_ogroup.inc