Bugfix for trac #8028
Fix a bug in filterLDAP::get_list which gets triggered, if the class is
used with multiple storage points and the search shall return objects in
the search base.
In this case the function would return much to many objects,
since multiple queries with scope LDAP_SCOPE_SUBTREE are issued, while
the query on the initial base should use LDAP_SCOPE_ONE instead.

This commit fixes this bug, by checking the base of the query against
the initial base and using a LDAP_SCOPE_ONE search, if initial base and
currently processed base are equal.