Bugfix in save logic for printers
Previously, when saving a user object after applying a template
the printer were tried to be added to the template UID. This
lead to an misleading error message and the printer beeing
unassigned from the template itself.

This change makes the save logic use the right uid for the printer
actions, by storing it in adapt_from_template and using the stored
value.