Add support for overlays to filter class
Add a new feature to the filter class, which allows to use
overlays in list classes, which make it possible to execute
queries regardless of the defined filter. This is useful for
example to include incoming objects in all bases and does
not require copy-paste-code to define the queries in each
and every query.
Beyond that a default filter currently can not have multiple
queries, so this also worksaround that fact.