Fix wrong sizelimit warning (#8319)
The group management used a very wide filter to determine which groups
have menu settings. Since menu groups exists in various releases, this
increases the number of results about several thousand times. This
triggers the sizelimit warning, which does not know about that
speciality. Apart from this, its probably performance critical, since
that gigantic list needs to be processed in PHP.

This commit fixes that problem, by simply issueing a single LDAP query
per group. Potentially this can degrade performance when there are
really much groups, but in the most cases, this will not be a problem
and even perform better.