Refactor tab initialization
There are several different ways in this module, where an object is
opened for editting. In all these cases its required to set FAIstate
and (eventually) read_only-mode on the tabObject. Previously this
was done independently in each code path, which lead to different
behaviour, depending on how the object is opened for editting.
This commit introduces a new function, which handles the attribute
initialisation and makes sure this function is called instead of
doing the initialisation manually.