Fix errornous list size in iceweasel/squeeze
Bugfix for Trac #8277
The javascript code for resizing of listings, when resizing the browser,
did change the wrong attributes, if using a squeeze version of
iceweasel. It seems condition blocks for IE vs. other browsers
were just swapped. This commit fixes this. Size for listings is now
right for firefox/iceweasel in the squeeze version and for chromium.