Make dn display an own line below the path display
Use a seperate div, below the header, to show the DN, because
otherwise long DNs break the layout.
For this change the genPathNavigator function had to be extended to
return an array with path and dnHeader. Also the genPathMenu function
had to be extended to use that second value and append it
to the string it returns.