Backport and adapt revision r21083
Revision 21083 fixes a reload bug in gotomasses that comes
from various bugs (syntax errors, trying to check form element, although
a div is not a form element etc.)
However, the code is somewhat inefficent, because it iterates over
all elements in the page. This commit merges the change in 21083, but
limits it to iterate over div elements.