Work-around #9173: Switch groupRDN and userRDN
This works around a problem when userRDN is empty (i.e. users are not
stored in a sub-ou).  If a subtree search is requested, only the
groupRDN location was searched via extensible match DNFilter search.  If
userRDN is put last, the groupRDN DNFilter gets cleared by the
filterLDAP class, resulting in a full subsearch.