Add simple CUPS URI check.
As CUPS URIs aren't RFC3986 complient, there is not much to check.
The scheme always matches the registered backends in
/usr/lib/cups/backend. These are all Unix filenames, which can
contain all characters except '/'.

This patch just checks for a valid RFC3986 scheme and the CUPS ':/'
seperator from the authority, in the hope for sane backend names.

Fixes Part 1 of TRAC #8778.