Fix detection of member deletes in group plugin
When a member is deleted from a group, the handle_removed_group_member_dns()
from the environment class is called in order to clean up shares.  However, the
logic of comparing the member arrays via array_diff() was returning incorrect
results if there was more than one member in the group.  Instead, extract the
DNs of the members and compare them directly.