added main/initrd_autopack for automatic initrd building Use this like this:
[tftp]
/initrd.img = |/usr/lib/go-susi/initrd_autopack /srv/fai/tftp/ /srv/fai/tftp/extracted

The first TFTP request for /initrd.img will cause it to be extracted to
/srv/fai/tftp/extracted/initrd.img/
and from that point on go-susi will automatically build a new initrd
whenever you make a change in that directory. To avoid the overhead of
building an initrd every time it is requested, initrd_autopack will
cache the initrd in /srv/fai/tftp/ and serve it directly from there
if ŧhe extracted/... directory has not changed.