Followup commit for LHM #10735: Bugfix
If an object with an xdg-spec is opened, the code detects that the saved
value is not an image and therefore does not try to display an image and
instead checks the "use the xdg spec"-checkbox. However, the code does
not set the iconPath attribute and therefore the xdg-specification
itself is not shown. This code changes that, by setting the value on
object construction.